pyproject.toml
[project]
name = "repo-analyzer"
version = "0.1.0"
description = "LangGraph-based GitHub repository analyzer"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.50",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "httpx>=0.27",
    "tavily-python>=0.5",
    "python-dotenv>=1.0",
    "tenacity>=9.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.21",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/repo_analyzer"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-m 'not live'"
markers = [
    "live: requires real API keys (run manually)",
]
